refactor: extract useMainConfig hook
